Update PrintPrimes.java
Changed calculatePrime method to be clearer and deleted to comment. It is common knowledge that 2 is the only even prime and the first prime number, so it does not need to be specifically mentioned as a comment. Instead, giving it the variable name firstPrime should be enough to make it clear.